Implement I2C watchdog to solve bus locks.
I2C multiplexer getting stuck during blade power up, which causes
the power sequence to shut the Zynq power off since it was not able
to read the boot status bit.

To acomplished that, an abstraction layer was created above the I2C
calls provided by the CERN IPMC framework, and  all the I2C calls were 
forwarded to that, which in turn make use of the low level I2C functions.
Doing that, a countdown counter was able to be decreased when any sensor
I2C bus transaction returned a problem. If the countdown reaches zero, 
it means that the sensor I2C bus was stuck for too long and a I2C 
multiplexer reset is issued.

On the other hand, any successful transacion on that bus reset the 
countdown counter.