Remove hadle-watch during power sequence.
Power sequence was able to be interrupted in the IPMC level when
watching for changes in the hot-swap handle. However, for instance,
When Zynq started its shutdown process, a power-cycle or a reet is 
needed to bring it back to work, which makes the power sequence
too complex.

The best way to solve this is to let the IPMC core to take care
about the handle. Once the board is shutted down, the IPMC will
restart the power-up sequence if the handle is closed.