Improve dynamic sensor handling.
Up to now, the implemented dynamic sensor handling was able to avoid
crashes when a sensor was not available. However, it kept the last
value seen for a sensor unless the sensor was not visible from the
beginning ("no reading").

With these changes, communication error causes the related sensor to
be disabled, and be listed as such. Successful communication enable a
sensor automatically.